+++ src/DonorDashboards/resources/js/app/components/donation-row/index.js
@@ -1,6 +1,7 @@
 import { Link } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 import { useWindowSize } from '../../hooks';
 
 const DonationRow = ( { donation } ) => {