+++ src/DonorDashboards/resources/js/app/components/subscription-cancel-modal/index.js
@@ -1,7 +1,8 @@
 import Button from '../button';
 import { cancelSubscriptionWithAPI } from './utils';
 
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 
 import './style.scss';