+++ src/DonorDashboards/resources/js/app/tabs/donation-history/dashboard-content.js
@@ -6,7 +6,8 @@ import Stats from './stats';
 
 import { useSelector } from './hooks';
 
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 
 const DashboardContent = () => {
 	const donations = useSelector( ( state ) => state.donations );