+++ src/DonorDashboards/resources/js/app/tabs/donation-history/stats/index.js
@@ -1,7 +1,8 @@
 import { useSelector } from '../hooks';
 import { useAccentColor } from '../../../hooks';
 
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 
 import './style.scss';
 import { Fragment } from 'react';