+++ src/DonorDashboards/resources/js/app/tabs/recurring-donations/content.js
@@ -1,7 +1,8 @@
 import { useLocation, Link } from 'react-router-dom';
 import { Fragment } from 'react';
 
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 
 import Heading from '../../components/heading';
 import Divider from '../../components/divider';