+++ src/DonorDashboards/App.php
@@ -116,7 +116,7 @@ class App {
 		wp_enqueue_script(
 			'give-donor-dashboards-app',
 			GIVE_PLUGIN_URL . 'assets/dist/js/donor-dashboards-app.js',
-			[ 'wp-element', 'wp-i18n' ],
+			[],
 			GIVE_VERSION,
 			true
 		);