+++ src/DonorDashboards/resources/js/app/components/tab-menu/index.js
@@ -2,7 +2,8 @@
 import { useState, Fragment } from 'react';
 import { useSelector } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 
 // Internal dependencies
 import TabLink from '../tab-link';