+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+-   Donor Dashboard UI is now consistent and polished (#5741)
+
 ## 2.10.0-rc.2 - 2021-03-18
 
 ### Fixed