+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2.10.0 - 2021-03-22
+
 ### Changed
 
 -   Donor Dashboard UI is now consistent and polished (#5741)